upload/requirements.txt
async-timeout==4.0.2
autopep8==1.7.0
boto3==1.24.89
botocore==1.27.89
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cryptography==38.0.2
DateTime==4.7
Deprecated==1.2.13
Flask==2.2.2
Flask-Limiter==2.6.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonify==0.5
limits==2.7.0
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.31
packaging==21.3
protobuf==3.20.1
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.13.0
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
pyuwsgi==2.0.20
requests==2.28.1
rich==12.6.0
s3transfer==0.6.0
six==1.16.0
toml==0.10.2
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
zope.interface==5.5.0
